# openspec/changes/sdlc-visualizer/interface-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arsitect SDLC Visualizer API
  version: 1.0.0
  description: Arsitect 可视化驾驶舱 REST API 契约（基于详细设计 20/21 模块导出）
servers:
- url: /api/v1
  description: API v1 基线
paths:
  /executions/trigger:
    post:
      tags:
      - skill-executions
      summary: Trigger Execution
      description: '触发 Skill 执行。


        根据触发动作创建新的 SkillExecution 记录。'
      operationId: trigger_execution_api_v1_executions_trigger_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionTriggerDTO'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillExecutionResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /executions/{execution_id}/status:
    get:
      tags:
      - skill-executions
      summary: Get Execution Status
      description: 查询执行状态。
      operationId: get_execution_status_api_v1_executions__execution_id__status_get
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: last_anchor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Anchor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStatusDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /executions/{execution_id}/logs:
    get:
      tags:
      - skill-executions
      summary: Get Execution Logs
      description: 查询执行日志。
      operationId: get_execution_logs_api_v1_executions__execution_id__logs_get
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: level
        in: query
        required: false
        schema:
          type: string
          default: ALL
          title: Level
      - name: anchor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Anchor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogQueryResultDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /executions/{execution_id}/retry:
    post:
      tags:
      - skill-executions
      summary: Retry Execution
      description: 重试失败的执行。
      operationId: retry_execution_api_v1_executions__execution_id__retry_post
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetryResultDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /executions/{execution_id}/confirm-release:
    post:
      tags:
      - skill-executions
      summary: Confirm Release
      description: 确认发布类 Skill 的执行。
      operationId: confirm_release_api_v1_executions__execution_id__confirm_release_post
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillExecutionResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /skills:
    get:
      tags:
      - skills
      summary: List Skills
      description: List registered skills with optional filters.
      operationId: list_skills_api_v1_skills_get
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: pattern
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pattern
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /skills/{skill_id}:
    get:
      tags:
      - skills
      summary: Get Skill
      description: Get a single skill by ID.
      operationId: get_skill_api_v1_skills__skill_id__get
      parameters:
      - name: skill_id
        in: path
        required: true
        schema:
          type: string
          title: Skill Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - skills
      summary: Delete Skill
      description: Delete a skill by ID.
      operationId: delete_skill_api_v1_skills__skill_id__delete
      parameters:
      - name: skill_id
        in: path
        required: true
        schema:
          type: string
          title: Skill Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ExecutionStatusDTO:
      properties:
        execution_id:
          type: string
          title: Execution Id
          description: 执行 ID
        current_phase:
          type: string
          title: Current Phase
          description: '当前阶段: PREP | EXEC | POST | NONE'
        phase_status:
          type: string
          title: Phase Status
          description: '阶段状态: RUNNING | COMPLETED | FAILED'
        overall_status:
          type: string
          title: Overall Status
          description: '整体状态: NOT_STARTED | RUNNING | SUCCESS | FAILED | UNKNOWN'
        stage_progress_percent:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Stage Progress Percent
          description: Stage 进度百分比
        status_timestamp:
          type: string
          format: date-time
          title: Status Timestamp
          description: 状态时间戳
        artifact_paths:
          items:
            type: string
          type: array
          title: Artifact Paths
          description: 产物路径列表
        error_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Summary
          description: 错误摘要
      type: object
      required:
      - execution_id
      - current_phase
      - phase_status
      - overall_status
      - stage_progress_percent
      - status_timestamp
      title: ExecutionStatusDTO
      description: 执行状态 DTO.
    ExecutionTriggerDTO:
      properties:
        trigger_action:
          type: string
          title: Trigger Action
          description: '触发动作: SINGLE_EXECUTE | BATCH_EXECUTE | RETRY'
        target_stage_id:
          type: string
          title: Target Stage Id
          description: 目标 Stage ID
        target_skill_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Skill Name
          description: SINGLE_EXECUTE 时必填
        confirm_release:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Confirm Release
          description: 发布类 Skill 必须 true
        previous_execution_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous Execution Id
          description: RETRY 时必填
      type: object
      required:
      - trigger_action
      - target_stage_id
      title: ExecutionTriggerDTO
      description: 执行触发请求 DTO.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LogEntryDTO:
      properties:
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: 时间戳
        level:
          type: string
          title: Level
          description: 级别
        content:
          type: string
          title: Content
          description: 内容
      type: object
      required:
      - timestamp
      - level
      - content
      title: LogEntryDTO
      description: 日志条目 DTO.
    LogQueryResultDTO:
      properties:
        log_entries:
          items:
            $ref: '#/components/schemas/LogEntryDTO'
          type: array
          title: Log Entries
          description: 日志条目
        total_count:
          type: integer
          minimum: 0.0
          title: Total Count
          description: 总条数
        next_anchor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Anchor
          description: 下一页游标
      type: object
      required:
      - total_count
      title: LogQueryResultDTO
      description: 日志查询结果 DTO.
...
    RetryResultDTO:
      properties:
        success:
          type: boolean
          title: Success
          description: 是否成功
        new_execution_id:
          anyOf:
          - type: string
          - type: 'null'
          title: New Execution Id
          description: 新执行 ID
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
          description: 消息
      type: object
      required:
      - success
      title: RetryResultDTO
      description: 重试结果 DTO.
...
    SkillExecutionResponseDTO:
      properties:
        execution_id:
          type: string
          title: Execution Id
          description: 执行 ID
        project_id:
          type: string
          title: Project Id
          description: 项目 ID
        stage_id:
          type: string
          title: Stage Id
          description: Stage ID
        skill_id:
          type: string
          title: Skill Id
          description: Skill ID
        skill_name:
          type: string
          title: Skill Name
          description: Skill 名称
        trigger_action:
          type: string
          title: Trigger Action
          description: 触发动作
        current_phase:
          type: string
          title: Current Phase
          description: 当前阶段
        phase_status:
          type: string
          title: Phase Status
          description: 阶段状态
        overall_status:
          type: string
          title: Overall Status
          description: 整体状态
        retry_count:
          type: integer
          title: Retry Count
          description: 重试次数
        previous_execution_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous Execution Id
          description: 前次执行 ID
        is_release_skill:
          type: boolean
          title: Is Release Skill
          description: 是否为发布类 Skill
        release_confirmed:
          type: boolean
          title: Release Confirmed
          description: 是否已确认发布
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
          description: 开始时间
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
          description: 完成时间
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 创建时间
      type: object
      required:
      - execution_id
      - project_id
      - stage_id
      - skill_id
      - skill_name
      - trigger_action
      - current_phase
      - phase_status
      - overall_status
      - retry_count
      - is_release_skill
      - release_confirmed
      - created_at
      title: SkillExecutionResponseDTO
      description: SkillExecution 响应 DTO.
...
    SkillListResponseDTO:
      properties:
        data:
          items:
            $ref: '#/components/schemas/SkillResponseDTO'
          type: array
          title: Data
        total_count:
          type: integer
          title: Total Count
      type: object
      required:
      - data
      - total_count
      title: SkillListResponseDTO
      description: Response model for skill list.
    SkillResponseDTO:
      properties:
        skill_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Id
        skill_name:
          type: string
          title: Skill Name
        version:
          type: string
          title: Version
        pattern:
          type: string
          title: Pattern
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        platforms:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Platforms
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        directory_path:
          type: string
          title: Directory Path
        parse_status:
          type: string
          title: Parse Status
          default: PARSED
      type: object
      required:
      - skill_name
      - version
      - pattern
      - directory_path
      title: SkillResponseDTO
      description: Response model for Skill data.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: MVP 阶段为本地会话占位，P1 升级为 JWT
...